Better .gitignore

+++ minitel-lib/src/main/kotlin/org/jraf/minitel/lib/util/color/Color.kt
@@ -1,3 +1,28 @@
+/*
+ * This source is part of the
+ *      _____  ___   ____
+ *  __ / / _ \/ _ | / __/___  _______ _
+ * / // / , _/ __ |/ _/_/ _ \/ __/ _ `/
+ * \___/_/|_/_/ |_/_/ (_)___/_/  \_, /
+ *                              /___/
+ * repository.
+ *
+ * Copyright (C) 2018 Benoit 'BoD' Lubek (BoD@JRAF.org)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package org.jraf.minitel.lib.util.color
 
 typealias AwtColor = java.awt.Color
@@ -10,7 +35,7 @@ fun rgbToHsl(color: AwtColor): Triple<Float, Float, Float> {
 fun rgbToHsl(r: Float, g: Float, b: Float): Triple<Float, Float, Float> {
     val max = Math.max(Math.max(r, g), b)
     val min = Math.min(Math.min(r, g), b)
-    val c = (max - min).toFloat()
+    val c = (max - min)
 
     var h_ = 0f
     if (c == 0f) {
